Move core modules to public folder
