+++ public/jugador.js
